+++ src/frontend.h
@@ -93,7 +93,7 @@ public:
                                   const ParseError &parseError) = 0;
 
     virtual bool handleTypeError(const std::vector<std::unique_ptr<FrontEnd::Context>> &ctx,
-                                 const TypeCheckError &checkError) = 0;
+                                 const TypeCheckError &checkError, bool firstAppear = true) = 0;
   };
 
   struct ModuleProvider {